deployment/requirements.txt
wordcloud==1.9.4
streamlit==1.40.1
seaborn==0.13.2
scikit-learn==1.5.2
numpy==1.26.4
nltk==3.9.1
matplotlib==3.9.2
joblib==1.4.2
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-api-core==2.23.0
googleapis-common-protos==1.66.0
python-dotenv==1.0.1